==> 2_inter_cazy_relation.ipynb <==
from transformers import AutoTokenizer, AutoModelForTokenClassification, pipeline
import spacy
import sys
import gpt3
import cross_coref
import entity_linkage

tokenizer = AutoTokenizer.from_pretrained("pucpr/clinicalnerpt-chemical")

model = AutoModelForTokenClassification.from_pretrained("pucpr/clinicalnerpt-chemical")

model_infer = pipeline('ner',model=model,tokenizer=tokenizer)

nlp = spacy.load("en_core_web_sm")

done = set()
is_header = False

for line in open("./raw_data/cazy_gpt3_extract.tsv", 'r'):
    if is_header == False:
        is_header = True
    else:
        done.add(line.split("\t")[0])

done

input_file = "./raw_data/cazy_activity.tsv"


with open(input_file, 'r', newline="", encoding='utf-8') as infile:

    for line in infile:
        fields = line.split("\t")
        cazy = fields[0]
        raw_text = fields[1]

        #if cazy == "CBM80":
        if cazy.startswith("CBM") and len(raw_text.strip()) > 0 and cazy not in done:
            #print (cazy, raw_text)
            sentences = (cross_coref.resolve_pronoun(raw_text))

            #print (sentences)

            #print (f'"{sent.text.strip()}"')
            res = gpt3.extract_relation("gpt3_training.txt", "#" + sentences.strip() + "\n").split("\n")
            #print("res:", cazy, raw_text, res)

            for r in res:
                results = r.split("|")
                #print (results)
                content = ""
                if len(results) == 2 and results[0] != "COEXISTS" and results[0] != "RELATES":
                    #print ("hello")
                    #print (f"{doi}|{is_primary}|{results[0]}|{results[1]}|{results[2]}")
                    #print ("results[1]", results[1])
                    disam = entity_linkage.name_disambiguation(results[1])
                    if disam != "":
                        results[1] = disam

                    content = f"{cazy}\t{sentences.strip()}\t{cazy}|" + "|".join(results) + "\n"
                else:
                    content = f"{cazy}\t{sentences.strip()}\t{cazy}|" + "|".join(results) + "\n"
                
                nonsensical_relates = False
                if results[0] == "RELATES":
                    if results[1] == cazy:
                        nonsensical_relates = True
                    if results[1] not in sentences:
                        nonsensical_relates = True

                if nonsensical_relates == False:
                    activity_output = open("./raw_data/cazy_gpt3_extract.tsv", 'a')
                    activity_output.write(content)
                    activity_output.close()
